Move collection, catalog, and item back to their own modules
