+++ pystac/asset.py
@@ -4,9 +4,9 @@ from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 from pystac import common_metadata, utils
 
 if TYPE_CHECKING:
+    from pystac.collection import Collection as Collection_Type
     from pystac.common_metadata import CommonMetadata as CommonMetadata_Type
-    from pystac.core import Collection as Collection_Type
-    from pystac.core import Item as Item_Type
+    from pystac.item import Item as Item_Type
 
 
 class Asset:
