+++ pystac/extensions/file.py
@@ -7,14 +7,16 @@ import enum
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 from pystac import asset as asset_mod
-from pystac import core, errors, stac_object, utils
+from pystac import errors
+from pystac import item as item_mod
+from pystac import stac_object, utils
 from pystac.extensions import base, hooks
 from pystac.serialization import identify
 
 if TYPE_CHECKING:
     from pystac.asset import Asset as Asset_Type
-    from pystac.core import Collection as Collection_Type
-    from pystac.core import Item as Item_Type
+    from pystac.collection import Collection as Collection_Type
+    from pystac.item import Item as Item_Type
     from pystac.serialization.identify import (
         STACJSONDescription as STACJSONDescription_Type,
     )
@@ -108,7 +110,7 @@ class FileExtension(
     def __init__(self, asset: "Asset_Type"):
         self.asset_href = asset.href
         self.properties = asset.extra_fields
-        if asset.owner and isinstance(asset.owner, core.Item):
+        if asset.owner and isinstance(asset.owner, item_mod.Item):
             self.additional_read_properties = [asset.owner.properties]
 
     def __repr__(self) -> str:
