+++ pystac/extensions/pointcloud.py
@@ -16,13 +16,15 @@ from typing import (
 )
 
 from pystac import asset as asset_mod
-from pystac import core, errors, stac_object, utils
+from pystac import errors
+from pystac import item as item_mod
+from pystac import stac_object, utils
 from pystac.extensions import base, hooks
 
 if TYPE_CHECKING:
     from pystac.asset import Asset as Asset_Type
-    from pystac.core import Collection as Collection_Type
-    from pystac.core import Item as Item_Type
+    from pystac.collection import Collection as Collection_Type
+    from pystac.item import Item as Item_Type
     from pystac.summaries import RangeSummary as RangeSummary_Type
 
 T = TypeVar("T", "Item_Type", "Asset_Type")
@@ -456,11 +458,11 @@ class PointcloudExtension(
 
             pystac.ExtensionTypeError : If an invalid object type is passed.
         """
-        if isinstance(obj, core.Item):
+        if isinstance(obj, item_mod.Item):
             cls.validate_has_extension(obj, add_if_missing)
             return cast(PointcloudExtension[T], ItemPointcloudExtension(obj))
         elif isinstance(obj, asset_mod.Asset):
-            if obj.owner is not None and not isinstance(obj.owner, core.Item):
+            if obj.owner is not None and not isinstance(obj.owner, item_mod.Item):
                 raise errors.ExtensionTypeError(
                     "Pointcloud extension does not apply to Collection Assets."
                 )
@@ -508,7 +510,7 @@ class AssetPointcloudExtension(PointcloudExtension["Asset_Type"]):
     def __init__(self, asset: "Asset_Type"):
         self.asset_href = asset.href
         self.properties = asset.extra_fields
-        if asset.owner and isinstance(asset.owner, core.Item):
+        if asset.owner and isinstance(asset.owner, item_mod.Item):
             self.additional_read_properties = [asset.owner.properties]
             self.repr_id = f"href={asset.href} item.id={asset.owner.id}"
         else:
