+++ pystac/extensions/projection.py
@@ -18,13 +18,15 @@ from typing import (
 )
 
 from pystac import asset as asset_mod
-from pystac import core, errors, stac_object
+from pystac import errors
+from pystac import item as item_mod
+from pystac import stac_object
 from pystac.extensions import base, hooks
 
 if TYPE_CHECKING:
     from pystac.asset import Asset as Asset_Type
-    from pystac.core import Collection as Collection_Type
-    from pystac.core import Item as Item_Type
+    from pystac.collection import Collection as Collection_Type
+    from pystac.item import Item as Item_Type
 
 T = TypeVar("T", "Item_Type", "Asset_Type")
 
@@ -283,7 +285,7 @@ class ProjectionExtension(
 
             pystac.ExtensionTypeError : If an invalid object type is passed.
         """
-        if isinstance(obj, core.Item):
+        if isinstance(obj, item_mod.Item):
             cls.validate_has_extension(obj, add_if_missing)
             return cast(ProjectionExtension[T], ItemProjectionExtension(obj))
         elif isinstance(obj, asset_mod.Asset):
@@ -348,7 +350,7 @@ class AssetProjectionExtension(ProjectionExtension["Asset_Type"]):
     def __init__(self, asset: "Asset_Type"):
         self.asset_href = asset.href
         self.properties = asset.extra_fields
-        if asset.owner and isinstance(asset.owner, core.Item):
+        if asset.owner and isinstance(asset.owner, item_mod.Item):
             self.additional_read_properties = [asset.owner.properties]
 
     def __repr__(self) -> str:
