+++ pystac/extensions/raster.py
@@ -7,13 +7,15 @@ import enum
 from typing import TYPE_CHECKING, Any, Dict, Iterable, List, Optional, Union
 
 from pystac import asset as asset_mod
-from pystac import core, errors, utils
+from pystac import errors
+from pystac import item as item_mod
+from pystac import utils
 from pystac.extensions import base
 
 if TYPE_CHECKING:
     from pystac.asset import Asset as Asset_Type
-    from pystac.core import Collection as Collection_Type
-    from pystac.core import Item as Item_Type
+    from pystac.collection import Collection as Collection_Type
+    from pystac.item import Item as Item_Type
 
 SCHEMA_URI = "https://stac-extensions.github.io/raster/v1.0.0/schema.json"
 
@@ -654,7 +656,7 @@ class RasterExtension(
     def __init__(self, asset: "Asset_Type"):
         self.asset_href = asset.href
         self.properties = asset.extra_fields
-        if asset.owner and isinstance(asset.owner, core.Item):
+        if asset.owner and isinstance(asset.owner, item_mod.Item):
             self.additional_read_properties = [asset.owner.properties]
 
     def __repr__(self) -> str:
