+++ pystac/extensions/scientific.py
@@ -22,12 +22,15 @@ from typing import (
 )
 from urllib import parse
 
-from pystac import core, errors, link, stac_object, utils
+from pystac import collection as collection_mod
+from pystac import errors
+from pystac import item as item_mod
+from pystac import link, stac_object, utils
 from pystac.extensions import base, hooks
 
 if TYPE_CHECKING:
-    from pystac.core import Collection as Collection_Type
-    from pystac.core import Item as Item_Type
+    from pystac.collection import Collection as Collection_Type
+    from pystac.item import Item as Item_Type
     from pystac.link import Link as Link_Type
     from pystac.stac_object import STACObject as STACObject_Type
 
@@ -244,10 +247,10 @@ class ScientificExtension(
 
             pystac.ExtensionTypeError : If an invalid object type is passed.
         """
-        if isinstance(obj, core.Collection):
+        if isinstance(obj, collection_mod.Collection):
             cls.validate_has_extension(obj, add_if_missing)
             return cast(ScientificExtension[T], CollectionScientificExtension(obj))
-        if isinstance(obj, core.Item):
+        if isinstance(obj, item_mod.Item):
             cls.validate_has_extension(obj, add_if_missing)
             return cast(ScientificExtension[T], ItemScientificExtension(obj))
         else:
