+++ pystac/summaries.py
@@ -23,7 +23,8 @@ else:
     from typing_extensions import Protocol
 
 if TYPE_CHECKING:
-    from pystac.core import Collection as Collection_Type, Item as Item_Type
+    from pystac.item import Item as Item_Type
+    from pystac.collection import Collection as Collection_Type
 
 from abc import abstractmethod
